# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Music Search Engine API
  description: |
    API for a personal music library web application for uploading, organizing,
    searching, and streaming legal music files.
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: https://api.example.com/api/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /tracks:
    get:
      tags: [Track]
      summary: List tracks
      operationId: listTracks
      parameters:
        - name: artist
          in: query
          schema:
            type: string
        - name: album
          in: query
          schema:
            type: string
        - name: genre
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/lastKey'
        - $ref: '#/components/parameters/sortBy'
        - $ref: '#/components/parameters/sortOrder'
      responses:
        '200':
          description: List of tracks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tracks/{id}:
    get:
      tags: [Track]
      summary: Get track details
      operationId: getTrack
      parameters:
        - $ref: '#/components/parameters/trackId'
      responses:
        '200':
          description: Track details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Track]
      summary: Update track metadata
      operationId: updateTrack
      parameters:
        - $ref: '#/components/parameters/trackId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrackRequest'
      responses:
        '200':
          description: Updated track
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Track]
      summary: Delete track
      operationId: deleteTrack
      parameters:
        - $ref: '#/components/parameters/trackId'
      responses:
        '204':
          description: Track deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /tracks/{id}/tags:
    post:
      tags: [Track, Tag]
      summary: Add tags to track
      operationId: addTagsToTrack
      parameters:
        - $ref: '#/components/parameters/trackId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTagsToTrackRequest'
      responses:
        '200':
          description: Track with added tags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /tracks/{id}/tags/{tag}:
    delete:
      tags: [Track, Tag]
      summary: Remove tag from track
      operationId: removeTagFromTrack
      parameters:
        - $ref: '#/components/parameters/trackId'
        - name: tag
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Track with tag removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /tags:
    get:
      tags: [Tag]
      summary: List tags
      operationId: listTags
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/lastKey'
        - $ref: '#/components/parameters/sortBy'
        - $ref: '#/components/parameters/sortOrder'
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Tag]
      summary: Create tag
      operationId: createTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagRequest'
      responses:
        '201':
          description: Created tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tags/{name}:
    put:
      tags: [Tag]
      summary: Update tag
      operationId: updateTag
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTagRequest'
      responses:
        '200':
          description: Updated tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Tag]
      summary: Delete tag
      operationId: deleteTag
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Tag deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /stream/{trackId}:
    get:
      tags: [Streaming]
      summary: Get streaming URL
      operationId: getStreamURL
      parameters:
        - $ref: '#/components/parameters/trackId'
        - name: quality
          in: query
          schema:
            type: string
            enum: [original, high, medium, low]
            default: original
      responses:
        '200':
          description: Streaming URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /download/{trackId}:
    get:
      tags: [Streaming]
      summary: Get download URL
      operationId: getDownloadURL
      parameters:
        - $ref: '#/components/parameters/trackId'
      responses:
        '200':
          description: Download URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /playback/record:
    post:
      tags: [Streaming]
      summary: Record playback event
      operationId: recordPlayback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordPlayRequest'
      responses:
        '204':
          description: Playback recorded
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /queue:
    get:
      tags: [Streaming]
      summary: Get play queue
      operationId: getQueue
      responses:
        '200':
          description: Play queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayQueue'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Streaming]
      summary: Update play queue
      operationId: updateQueue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQueueRequest'
      responses:
        '200':
          description: Updated play queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayQueue'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /queue/action:
    post:
      tags: [Streaming]
      summary: Perform queue action
      operationId: queueAction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueueActionRequest'
      responses:
        '200':
          description: Updated play queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayQueue'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    trackId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    lastKey:
      name: lastKey
      in: query
      schema:
        type: string
    sortBy:
      name: sortBy
      in: query
      schema:
        type: string
    sortOrder:
      name: sortOrder
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        lastKey:
          type: string
        nextKey:
          type: string
        totalEstimate:
          type: integer
...
    TrackResponse:
      type: object
      required: [id, title, artist, duration, durationStr, format, fileSize, fileSizeStr, playCount, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        artist:
          type: string
        albumArtist:
          type: string
        album:
          type: string
        albumId:
          type: string
          format: uuid
        genre:
          type: string
        year:
          type: integer
        trackNumber:
          type: integer
        discNumber:
          type: integer
        duration:
          type: integer
        durationStr:
          type: string
        format:
          type: string
        fileSize:
          type: integer
          format: int64
        fileSizeStr:
          type: string
        coverArtUrl:
          type: string
          format: uri
        playCount:
          type: integer
        lastPlayed:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateTrackRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        artist:
          type: string
          minLength: 1
          maxLength: 500
        albumArtist:
          type: string
          maxLength: 500
        album:
          type: string
          maxLength: 500
        genre:
          type: string
          maxLength: 100
        year:
          type: integer
          minimum: 1
          maximum: 9999
        trackNumber:
          type: integer
          minimum: 0
        discNumber:
          type: integer
          minimum: 0
        lyrics:
          type: string
        comment:
          type: string
          maxLength: 1000
        tags:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 50
    AddTagsToTrackRequest:
      type: object
      required: [tags]
      properties:
        tags:
          type: array
          minItems: 1
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 50
    TrackListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TrackResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    TagResponse:
      type: object
      required: [name, trackCount, createdAt, updatedAt]
      properties:
        name:
          type: string
        color:
          type: string
        trackCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateTagRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        color:
          type: string
          pattern: '^#[0-9a-fA-F]{6}$'
    UpdateTagRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        color:
          type: string
          pattern: '^#[0-9a-fA-F]{6}$'
    TagListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TagResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    StreamResponse:
      type: object
      required: [trackId, streamUrl, expiresAt, format]
      properties:
        trackId:
          type: string
          format: uuid
        streamUrl:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
        format:
          type: string
        bitrate:
          type: integer
    DownloadResponse:
      type: object
      required: [trackId, downloadUrl, expiresAt, fileName, fileSize, format]
      properties:
        trackId:
          type: string
          format: uuid
        downloadUrl:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
        fileName:
          type: string
        fileSize:
          type: integer
          format: int64
        format:
          type: string
    RecordPlayRequest:
      type: object
      required: [trackId, duration]
      properties:
        trackId:
          type: string
          format: uuid
        duration:
          type: integer
          minimum: 1
        completed:
          type: boolean
    PlayQueue:
      type: object
      required: [userId, currentIndex, trackIds, shuffleMode, repeatMode, updatedAt]
      properties:
        userId:
          type: string
        currentIndex:
          type: integer
        trackIds:
          type: array
          items:
            type: string
            format: uuid
        shuffleMode:
          type: boolean
        repeatMode:
          type: string
          enum: [none, one, all]
        updatedAt:
          type: string
          format: date-time
    UpdateQueueRequest:
      type: object
      properties:
        trackIds:
          type: array
          items:
            type: string
            format: uuid
        currentIndex:
          type: integer
        shuffleMode:
          type: boolean
        repeatMode:
          type: string
          enum: [none, one, all]
    QueueActionRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [add_next, add_last, remove, clear, shuffle, unshuffle]
        trackIds:
          type: array
          items:
            type: string
            format: uuid